move checkJST to utils/jwt
